+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 [![Downloads](https://img.shields.io/github/downloads/zevnda/steam-game-idler/1.7.2/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/zevnda/steam-game-idler/releases/download/1.7.2/Steam.Game.Idler_1.7.2_x64_en-US.msi)
 
 ## Changelog
-- Fixed some changelog modal styles
-- Update system icon to match UI
-  - Idle windows have a blue variant of the icon
+- New icons, again 😬
+- Replaced the icon in the `header` for a drag region
+- Added an `export data` button to `settings`
+  - Useful for debugging issues